attestation: re-factor directory structure
