+++ src/enclave/outside/attestation/CMakeLists.txt
@@ -8,8 +8,6 @@ set(SGX_DCAP_LIB /usr/lib/libsgx_dcap_ql.so)
 
 add_library(attestation STATIC
     attestation.cpp
-    generateQuote.cpp
-    validateQuote.cpp
     AzureAttestationServiceClient.cpp
     EnclaveInfo.cpp
 )
